pyproject.toml
[project]
name = "hcscraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "playwright>=1.49.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
